+++ R/protocol.R
@@ -133,5 +133,6 @@ MessageType <- list(
     Error = 1,
     Warning = 2,
     Info = 3,
-    Log = 4
+    Log = 4,
+    Debug = 5
 )